Hide fake instance creation in GUI
