+++ GUI/Dialogs/ManageGameInstancesDialog.resx
@@ -115,7 +115,7 @@
   <data name="SelectButton.Text" xml:space="preserve"><value>Select</value></data>
   <data name="AddNewButton.Text" xml:space="preserve"><value>New game instance</value></data>
   <data name="AddToCKANMenuItem.Text" xml:space="preserve"><value>Add instance to CKAN</value></data>
-  <data name="CloneFakeInstanceMenuItem.Text" xml:space="preserve"><value>Clone or fake new instance</value></data>
+  <data name="CloneGameInstanceMenuItem.Text" xml:space="preserve"><value>Clone instance</value></data>
   <data name="RenameButton.Text" xml:space="preserve"><value>Rename</value></data>
   <data name="SetAsDefaultCheckbox.Text" xml:space="preserve"><value>Set as default</value></data>
   <data name="ForgetButton.Text" xml:space="preserve"><value>Forget</value></data>
