+++ GUI/Localization/de-DE/CloneGameInstanceDialog.de-DE.resx
@@ -59,46 +59,46 @@
             : using a System.ComponentModel.TypeConverter
             : and then encoded with base64 encoding.
     -->
-  <xsd:schema xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata" id="root">
-    <xsd:import namespace="http://www.w3.org/XML/1998/namespace"/>
+  <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
+    <xsd:import namespace="http://www.w3.org/XML/1998/namespace" />
     <xsd:element name="root" msdata:IsDataSet="true">
       <xsd:complexType>
         <xsd:choice maxOccurs="unbounded">
           <xsd:element name="metadata">
             <xsd:complexType>
               <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0"/>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" />
               </xsd:sequence>
-              <xsd:attribute name="name" use="required" type="xsd:string"/>
-              <xsd:attribute name="type" type="xsd:string"/>
-              <xsd:attribute name="mimetype" type="xsd:string"/>
-              <xsd:attribute ref="xml:space"/>
+              <xsd:attribute name="name" use="required" type="xsd:string" />
+              <xsd:attribute name="type" type="xsd:string" />
+              <xsd:attribute name="mimetype" type="xsd:string" />
+              <xsd:attribute ref="xml:space" />
             </xsd:complexType>
           </xsd:element>
           <xsd:element name="assembly">
             <xsd:complexType>
-              <xsd:attribute name="alias" type="xsd:string"/>
-              <xsd:attribute name="name" type="xsd:string"/>
+              <xsd:attribute name="alias" type="xsd:string" />
+              <xsd:attribute name="name" type="xsd:string" />
             </xsd:complexType>
           </xsd:element>
           <xsd:element name="data">
             <xsd:complexType>
               <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1"/>
-                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2"/>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2" />
               </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1"/>
-              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3"/>
-              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4"/>
-              <xsd:attribute ref="xml:space"/>
+              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1" />
+              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3" />
+              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4" />
+              <xsd:attribute ref="xml:space" />
             </xsd:complexType>
           </xsd:element>
           <xsd:element name="resheader">
             <xsd:complexType>
               <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1"/>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
               </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required"/>
+              <xsd:attribute name="name" type="xsd:string" use="required" />
             </xsd:complexType>
           </xsd:element>
         </xsd:choice>
@@ -117,55 +117,15 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="radioButtonClone.Text" xml:space="preserve">
-    <value>Clonar instância existente do KSP</value>
-  </data>
-  <data name="labelOldInstance.Text" xml:space="preserve">
-    <value>Instância para clonar:</value>
-  </data>
-  <data name="labelOldPath.Text" xml:space="preserve">
-    <value>Casminho para clonar:</value>
-  </data>
-  <data name="buttonInstancePathSelection.Text" xml:space="preserve">
-    <value>Selecionar...</value>
-  </data>
-  <data name="radioButtonFake.Text" xml:space="preserve">
-    <value>Falsificar nova instância do KSP</value>
-  </data>
-  <data name="labelVersion.Text" xml:space="preserve">
-    <value>Versão:</value>
-  </data>
-  <data name="labelDlcVersions.Text" xml:space="preserve">
-    <value>Versões das DLC (vazio para nenhuma):</value>
-  </data>
-  <data name="labelMakingHistory.Text" xml:space="preserve">
-    <value>DLC Making History</value>
-  </data>
-  <data name="labelBreakingGround.Text" xml:space="preserve">
-    <value>DLC Breaking Ground</value>
-  </data>
-  <data name="labelNewName.Text" xml:space="preserve">
-    <value>Nome da nova instância:</value>
-  </data>
-  <data name="labelNewPath.Text" xml:space="preserve">
-    <value>Caminho da nova instância:</value>
-  </data>
-  <data name="buttonPathBrowser.Text" xml:space="preserve">
-    <value>Selecionar...</value>
-  </data>
-  <data name="checkBoxSetAsDefault.Text" xml:space="preserve">
-    <value>Definir a nova instância como padrão</value>
-  </data>
-  <data name="checkBoxSwitchInstance.Text" xml:space="preserve">
-    <value>Trocar para a nova instância</value>
-  </data>
-  <data name="buttonOK.Text" xml:space="preserve">
-    <value>Criar</value>
-  </data>
-  <data name="buttonCancel.Text" xml:space="preserve">
-    <value>Cancelar</value>
-  </data>
-  <data name="$this.Text" xml:space="preserve">
-    <value>Clonar ou Falsificar instância do KSP</value>
-  </data>
+  <data name="labelOldInstance.Text" xml:space="preserve"><value>Zu klonende Instanz:</value></data>
+  <data name="labelOldPath.Text" xml:space="preserve"><value>Pfad zur zu klonenden Instanz:</value></data>
+  <data name="buttonInstancePathSelection.Text" xml:space="preserve"><value>Suchen</value></data>
+  <data name="labelNewName.Text" xml:space="preserve"><value>Name der neuen Instanz:</value></data>
+  <data name="labelNewPath.Text" xml:space="preserve"><value>Pfad für die neue Instanz:</value></data>
+  <data name="buttonPathBrowser.Text" xml:space="preserve"><value>Suchen</value></data>
+  <data name="checkBoxSetAsDefault.Text" xml:space="preserve"><value>Als Standardinstanz festlegen</value></data>
+  <data name="checkBoxSwitchInstance.Text" xml:space="preserve"><value>Zur neuen Instanz wechseln</value></data>
+  <data name="buttonOK.Text" xml:space="preserve"><value>Erstellen</value></data>
+  <data name="buttonCancel.Text" xml:space="preserve"><value>Abbrechen</value></data>
+  <data name="$this.Text" xml:space="preserve"><value>Neue Instanz</value></data>
 </root>
