+++ GUI/Localization/de-DE/ManageGameInstancesDialog.de-DE.resx
@@ -125,7 +125,7 @@
   <data name="SelectButton.Text" xml:space="preserve"><value>Wählen</value></data>
   <data name="AddNewButton.Text" xml:space="preserve"><value>Neue Spielinstanz</value></data>
   <data name="AddToCKANMenuItem.Text" xml:space="preserve"><value>Instanz hinzufügen</value></data>
-  <data name="CloneFakeInstanceMenuItem.Text" xml:space="preserve"><value>Instanz klonen oder fälschen</value></data>
+  <data name="CloneGameInstanceMenuItem.Text" xml:space="preserve"><value>Instanz klonen</value></data>
   <data name="RenameButton.Text" xml:space="preserve"><value>Umbenennen</value></data>
   <data name="SetAsDefaultCheckbox.Text" xml:space="preserve"><value>Als Standardinstanz setzen</value></data>
   <data name="ForgetButton.Text" xml:space="preserve"><value>Vergessen</value></data>
