+++ GUI/Localization/fr-FR/CloneGameInstanceDialog.fr-FR.resx
@@ -59,46 +59,46 @@
             : using a System.ComponentModel.TypeConverter
             : and then encoded with base64 encoding.
     -->
-  <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
-    <xsd:import namespace="http://www.w3.org/XML/1998/namespace" />
+  <xsd:schema xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata" id="root">
+    <xsd:import namespace="http://www.w3.org/XML/1998/namespace"/>
     <xsd:element name="root" msdata:IsDataSet="true">
       <xsd:complexType>
         <xsd:choice maxOccurs="unbounded">
           <xsd:element name="metadata">
             <xsd:complexType>
               <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" />
+                <xsd:element name="value" type="xsd:string" minOccurs="0"/>
               </xsd:sequence>
-              <xsd:attribute name="name" use="required" type="xsd:string" />
-              <xsd:attribute name="type" type="xsd:string" />
-              <xsd:attribute name="mimetype" type="xsd:string" />
-              <xsd:attribute ref="xml:space" />
+              <xsd:attribute name="name" use="required" type="xsd:string"/>
+              <xsd:attribute name="type" type="xsd:string"/>
+              <xsd:attribute name="mimetype" type="xsd:string"/>
+              <xsd:attribute ref="xml:space"/>
             </xsd:complexType>
           </xsd:element>
           <xsd:element name="assembly">
             <xsd:complexType>
-              <xsd:attribute name="alias" type="xsd:string" />
-              <xsd:attribute name="name" type="xsd:string" />
+              <xsd:attribute name="alias" type="xsd:string"/>
+              <xsd:attribute name="name" type="xsd:string"/>
             </xsd:complexType>
           </xsd:element>
           <xsd:element name="data">
             <xsd:complexType>
               <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
-                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2" />
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1"/>
+                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2"/>
               </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1" />
-              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3" />
-              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4" />
-              <xsd:attribute ref="xml:space" />
+              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1"/>
+              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3"/>
+              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4"/>
+              <xsd:attribute ref="xml:space"/>
             </xsd:complexType>
           </xsd:element>
           <xsd:element name="resheader">
             <xsd:complexType>
               <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1"/>
               </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required" />
+              <xsd:attribute name="name" type="xsd:string" use="required"/>
             </xsd:complexType>
           </xsd:element>
         </xsd:choice>
@@ -117,21 +117,37 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="radioButtonClone.Text" xml:space="preserve"><value>克隆现存的KSP实例</value></data>
-  <data name="labelOldInstance.Text" xml:space="preserve"><value>克隆的实例</value></data>
-  <data name="labelOldPath.Text" xml:space="preserve"><value>克隆的路径</value></data>
-  <data name="buttonInstancePathSelection.Text" xml:space="preserve"><value>选择...</value></data>
-  <data name="radioButtonFake.Text" xml:space="preserve"><value>伪装新的KSP实例</value></data>
-  <data name="labelVersion.Text" xml:space="preserve"><value>版本:</value></data>
-  <data name="labelDlcVersions.Text" xml:space="preserve"><value>DLC版本（留空以忽略）:</value></data>
-  <data name="labelMakingHistory.Text" xml:space="preserve"><value>Making History</value></data>
-  <data name="labelBreakingGround.Text" xml:space="preserve"><value>Breaking Ground</value></data>
-  <data name="labelNewName.Text" xml:space="preserve"><value>新实例名称:</value></data>
-  <data name="labelNewPath.Text" xml:space="preserve"><value>新实例路径:</value></data>
-  <data name="buttonPathBrowser.Text" xml:space="preserve"><value>选择...</value></data>
-  <data name="checkBoxSetAsDefault.Text" xml:space="preserve"><value>将新实例设为默认</value></data>
-  <data name="checkBoxSwitchInstance.Text" xml:space="preserve"><value>切换到新实例</value></data>
-  <data name="buttonOK.Text" xml:space="preserve"><value>创建</value></data>
-  <data name="buttonCancel.Text" xml:space="preserve"><value>取消</value></data>
-  <data name="$this.Text" xml:space="preserve"><value>克隆或伪装KSP实例</value></data>
+  <data name="labelOldInstance.Text" xml:space="preserve">
+    <value>Instance à cloner :</value>
+  </data>
+  <data name="labelOldPath.Text" xml:space="preserve">
+    <value>Chemin à cloner :</value>
+  </data>
+  <data name="buttonInstancePathSelection.Text" xml:space="preserve">
+    <value>Parcourir</value>
+  </data>
+  <data name="labelNewName.Text" xml:space="preserve">
+    <value>Nom de la nouvelle instance :</value>
+  </data>
+  <data name="labelNewPath.Text" xml:space="preserve">
+    <value>Chemin de la nouvelle instance :</value>
+  </data>
+  <data name="buttonPathBrowser.Text" xml:space="preserve">
+    <value>Parcourir</value>
+  </data>
+  <data name="checkBoxSetAsDefault.Text" xml:space="preserve">
+    <value>Marquer par défaut</value>
+  </data>
+  <data name="checkBoxSwitchInstance.Text" xml:space="preserve">
+    <value>Basculer à la nouvelle instance</value>
+  </data>
+  <data name="buttonOK.Text" xml:space="preserve">
+    <value>Créer</value>
+  </data>
+  <data name="buttonCancel.Text" xml:space="preserve">
+    <value>Annuler</value>
+  </data>
+  <data name="$this.Text" xml:space="preserve">
+    <value>Cloner une Instance</value>
+  </data>
 </root>
