+++ GUI/Localization/ja-JP/ManageGameInstancesDialog.ja-JP.resx
@@ -114,7 +114,7 @@
   <data name="SelectButton.Text" xml:space="preserve"><value>選択</value></data>
   <data name="AddNewButton.Text" xml:space="preserve"><value>新規インスタンス</value></data>
   <data name="AddToCKANMenuItem.Text" xml:space="preserve"><value>CKANにインスタンスを追加</value></data>
-  <data name="CloneFakeInstanceMenuItem.Text" xml:space="preserve"><value>新規インスタンスを複製または作成</value></data>
+  <data name="CloneGameInstanceMenuItem.Text" xml:space="preserve"><value>クローンインスタンス</value></data>
   <data name="RenameButton.Text" xml:space="preserve"><value>名前変更</value></data>
   <data name="SetAsDefaultCheckbox.Text" xml:space="preserve"><value>デフォルトに設定</value></data>
   <data name="ForgetButton.Text" xml:space="preserve"><value>除去</value></data>
