+++ GUI/Localization/ru-RU/CloneGameInstanceDialog.ru-RU.resx
@@ -59,46 +59,46 @@
             : using a System.ComponentModel.TypeConverter
             : and then encoded with base64 encoding.
     -->
-  <xsd:schema xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata" id="root">
-    <xsd:import namespace="http://www.w3.org/XML/1998/namespace"/>
+  <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
+    <xsd:import namespace="http://www.w3.org/XML/1998/namespace" />
     <xsd:element name="root" msdata:IsDataSet="true">
       <xsd:complexType>
         <xsd:choice maxOccurs="unbounded">
           <xsd:element name="metadata">
             <xsd:complexType>
               <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0"/>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" />
               </xsd:sequence>
-              <xsd:attribute name="name" use="required" type="xsd:string"/>
-              <xsd:attribute name="type" type="xsd:string"/>
-              <xsd:attribute name="mimetype" type="xsd:string"/>
-              <xsd:attribute ref="xml:space"/>
+              <xsd:attribute name="name" use="required" type="xsd:string" />
+              <xsd:attribute name="type" type="xsd:string" />
+              <xsd:attribute name="mimetype" type="xsd:string" />
+              <xsd:attribute ref="xml:space" />
             </xsd:complexType>
           </xsd:element>
           <xsd:element name="assembly">
             <xsd:complexType>
-              <xsd:attribute name="alias" type="xsd:string"/>
-              <xsd:attribute name="name" type="xsd:string"/>
+              <xsd:attribute name="alias" type="xsd:string" />
+              <xsd:attribute name="name" type="xsd:string" />
             </xsd:complexType>
           </xsd:element>
           <xsd:element name="data">
             <xsd:complexType>
               <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1"/>
-                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2"/>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
+                <xsd:element name="comment" type="xsd:string" minOccurs="0" msdata:Ordinal="2" />
               </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1"/>
-              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3"/>
-              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4"/>
-              <xsd:attribute ref="xml:space"/>
+              <xsd:attribute name="name" type="xsd:string" use="required" msdata:Ordinal="1" />
+              <xsd:attribute name="type" type="xsd:string" msdata:Ordinal="3" />
+              <xsd:attribute name="mimetype" type="xsd:string" msdata:Ordinal="4" />
+              <xsd:attribute ref="xml:space" />
             </xsd:complexType>
           </xsd:element>
           <xsd:element name="resheader">
             <xsd:complexType>
               <xsd:sequence>
-                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1"/>
+                <xsd:element name="value" type="xsd:string" minOccurs="0" msdata:Ordinal="1" />
               </xsd:sequence>
-              <xsd:attribute name="name" type="xsd:string" use="required"/>
+              <xsd:attribute name="name" type="xsd:string" use="required" />
             </xsd:complexType>
           </xsd:element>
         </xsd:choice>
@@ -117,55 +117,15 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="radioButtonClone.Text" xml:space="preserve">
-    <value>Clona istanza KSP esistente</value>
-  </data>
-  <data name="labelOldInstance.Text" xml:space="preserve">
-    <value>Istanza da clonare:</value>
-  </data>
-  <data name="labelOldPath.Text" xml:space="preserve">
-    <value>Percorso della clonazione:</value>
-  </data>
-  <data name="buttonInstancePathSelection.Text" xml:space="preserve">
-    <value>Seleziona...</value>
-  </data>
-  <data name="radioButtonFake.Text" xml:space="preserve">
-    <value>Falsificare una nuova istanza di KSP</value>
-  </data>
-  <data name="labelVersion.Text" xml:space="preserve">
-    <value>Versione:</value>
-  </data>
-  <data name="labelDlcVersions.Text" xml:space="preserve">
-    <value>Versione DLC (vuota per nessuna):</value>
-  </data>
-  <data name="labelMakingHistory.Text" xml:space="preserve">
-    <value>Making History</value>
-  </data>
-  <data name="labelBreakingGround.Text" xml:space="preserve">
-    <value>Breaking Ground</value>
-  </data>
-  <data name="labelNewName.Text" xml:space="preserve">
-    <value>Nome per la nuova istanza:</value>
-  </data>
-  <data name="labelNewPath.Text" xml:space="preserve">
-    <value>Percorso per la nuova istanza:</value>
-  </data>
-  <data name="buttonPathBrowser.Text" xml:space="preserve">
-    <value>Seleziona...</value>
-  </data>
-  <data name="checkBoxSetAsDefault.Text" xml:space="preserve">
-    <value>Imposta nuova istanza come predefinita</value>
-  </data>
-  <data name="checkBoxSwitchInstance.Text" xml:space="preserve">
-    <value>Passa alla nuova istanza</value>
-  </data>
-  <data name="buttonOK.Text" xml:space="preserve">
-    <value>Crea</value>
-  </data>
-  <data name="buttonCancel.Text" xml:space="preserve">
-    <value>Annulla</value>
-  </data>
-  <data name="$this.Text" xml:space="preserve">
-    <value>Clonare o falsificare un'istanza di KSP</value>
-  </data>
+  <data name="labelOldInstance.Text" xml:space="preserve"><value>Исходная сборка:</value></data>
+  <data name="labelOldPath.Text" xml:space="preserve"><value>Путь к папке сборки:</value></data>
+  <data name="buttonInstancePathSelection.Text" xml:space="preserve"><value>Выбрать...</value></data>
+  <data name="labelNewName.Text" xml:space="preserve"><value>Название новой сборки:</value></data>
+  <data name="labelNewPath.Text" xml:space="preserve"><value>Путь к новой сборке:</value></data>
+  <data name="buttonPathBrowser.Text" xml:space="preserve"><value>Выбрать...</value></data>
+  <data name="checkBoxSetAsDefault.Text" xml:space="preserve"><value>Назначить по умолчанию</value></data>
+  <data name="checkBoxSwitchInstance.Text" xml:space="preserve"><value>Переключиться на новую сборку</value></data>
+  <data name="buttonOK.Text" xml:space="preserve"><value>Создать</value></data>
+  <data name="buttonCancel.Text" xml:space="preserve"><value>Отмена</value></data>
+  <data name="$this.Text" xml:space="preserve"><value>Клонировать сборку</value></data>
 </root>
