+++ GUI/Localization/ru-RU/ManageGameInstancesDialog.ru-RU.resx
@@ -114,7 +114,7 @@
   <data name="SelectButton.Text" xml:space="preserve"><value>Выбрать</value></data>
   <data name="AddNewButton.Text" xml:space="preserve"><value>Добавить</value></data>
   <data name="AddToCKANMenuItem.Text" xml:space="preserve"><value>Добавить папку сборки в CKAN</value></data>
-  <data name="CloneFakeInstanceMenuItem.Text" xml:space="preserve"><value>Клонировать или имитировать сборку</value></data>
+  <data name="CloneGameInstanceMenuItem.Text" xml:space="preserve"><value>Клонировать экземпляр</value></data>
   <data name="RenameButton.Text" xml:space="preserve"><value>Переименовать</value></data>
   <data name="SetAsDefaultCheckbox.Text" xml:space="preserve"><value>По умолчанию</value></data>
   <data name="ForgetButton.Text" xml:space="preserve"><value>Забыть</value></data>
