+++ GUI/Localization/zh-CN/CloneGameInstanceDialog.zh-CN.resx
@@ -117,21 +117,15 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="radioButtonClone.Text" xml:space="preserve"><value>Clone existing KSP instance</value></data>
-  <data name="labelOldInstance.Text" xml:space="preserve"><value>Instance to clone:</value></data>
-  <data name="labelOldPath.Text" xml:space="preserve"><value>Path to clone:</value></data>
-  <data name="buttonInstancePathSelection.Text" xml:space="preserve"><value>Select...</value></data>
-  <data name="radioButtonFake.Text" xml:space="preserve"><value>Fake new KSP instance</value></data>
-  <data name="labelVersion.Text" xml:space="preserve"><value>Version:</value></data>
-  <data name="labelDlcVersions.Text" xml:space="preserve"><value>DLC versions (empty for none):</value></data>
-  <data name="labelMakingHistory.Text" xml:space="preserve"><value>Making History</value></data>
-  <data name="labelBreakingGround.Text" xml:space="preserve"><value>Breaking Ground</value></data>
-  <data name="labelNewName.Text" xml:space="preserve"><value>Name for the new instance:</value></data>
-  <data name="labelNewPath.Text" xml:space="preserve"><value>Path for the new instance:</value></data>
-  <data name="buttonPathBrowser.Text" xml:space="preserve"><value>Select...</value></data>
-  <data name="checkBoxSetAsDefault.Text" xml:space="preserve"><value>Set new instance as default</value></data>
-  <data name="checkBoxSwitchInstance.Text" xml:space="preserve"><value>Switch to new instance</value></data>
-  <data name="buttonOK.Text" xml:space="preserve"><value>Create</value></data>
-  <data name="buttonCancel.Text" xml:space="preserve"><value>Cancel</value></data>
-  <data name="$this.Text" xml:space="preserve"><value>Clone or Fake KSP Instance</value></data>
+  <data name="labelOldInstance.Text" xml:space="preserve"><value>克隆的实例</value></data>
+  <data name="labelOldPath.Text" xml:space="preserve"><value>克隆的路径</value></data>
+  <data name="buttonInstancePathSelection.Text" xml:space="preserve"><value>选择...</value></data>
+  <data name="labelNewName.Text" xml:space="preserve"><value>新实例名称:</value></data>
+  <data name="labelNewPath.Text" xml:space="preserve"><value>新实例路径:</value></data>
+  <data name="buttonPathBrowser.Text" xml:space="preserve"><value>选择...</value></data>
+  <data name="checkBoxSetAsDefault.Text" xml:space="preserve"><value>将新实例设为默认</value></data>
+  <data name="checkBoxSwitchInstance.Text" xml:space="preserve"><value>切换到新实例</value></data>
+  <data name="buttonOK.Text" xml:space="preserve"><value>创建</value></data>
+  <data name="buttonCancel.Text" xml:space="preserve"><value>取消</value></data>
+  <data name="$this.Text" xml:space="preserve"><value>克隆或伪装KSP实例</value></data>
 </root>
