+++ GUI/Properties/Resources.ko-KR.resx
@@ -124,14 +124,12 @@
   <data name="CloneFakeKspDialogEnterName" xml:space="preserve"><value>새로운 인스턴스의 이름을 입력해주세요.</value></data>
   <data name="CloneFakeKspDialogEnterPath" xml:space="preserve"><value>새로운 인스턴스의 경로를 입력해주세요.</value></data>
   <data name="CloneFakeKspDialogCloningInstance" xml:space="preserve"><value>"인스턴스 복제 중..."</value></data>
-  <data name="CloneFakeKspDialogDlcVersionMalformatted" xml:space="preserve"><value>{0} DLC의 버전이 올바르지 않아요. 예를 들어 "1.0.0"를 시도해보세요.</value></data>
   <data name="CloneFakeKspDialogInstanceNotValid" xml:space="preserve"><value>복사할 인스턴스가 유효하지 않아요. {0}</value></data>
   <data name="CloneFakeKspDialogDestinationNotEmpty" xml:space="preserve"><value>설정한 폴더가 비어있지 않아요. {0}</value></data>
   <data name="CloneFakeKspDialogCloneFailed" xml:space="preserve"><value>복사 실패. {0}</value></data>
   <data name="CloneFakeKspDialogSuccessfulClone" xml:space="preserve"><value>성공적으로 인스턴스를 복사했어요.</value></data>
   <data name="CloneFakeKspDialogCreatingInstance" xml:space="preserve"><value>새로운 인스턴스 생성중...</value></data>
   <data name="CloneFakeKspDialogNameAlreadyUsed" xml:space="preserve"><value>그 이름은 이미 사용 중이에요.</value></data>
-  <data name="CloneFakeKspDialogFakeFailed" xml:space="preserve"><value>인스턴스 위조 실패. {0}</value></data>
   <data name="CloneFakeKspDialogSuccessfulCreate" xml:space="preserve"><value>성공적으로 인스턴스를 생성했어요.</value></data>
   <data name="CompatibleGameVersionsDialogNone" xml:space="preserve"><value>&lt;NONE&gt;</value></data>
   <data name="CompatibleGameVersionsDialogGameUpdated" xml:space="preserve"><value>마지막으로 호환되는 버전을 검토한 뒤로 게임이 업데이트 되었어요. 설정이 정확한지 확인해주세요.</value></data>
